BenchMark/requirements.txt
absl-py==0.12.0
amqp==5.0.5
attrs==20.3.0
billiard==3.6.3.0
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
cirq==0.10.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cycler==0.10.0
decorator==4.4.2
deprecation==2.1.0
flatbuffers==1.12
google-api-core==1.26.2
google-auth==1.28.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
h5py==3.2.1
idna==2.10
iniconfig==1.1.1
ipython-genutils==0.2.0
jax==0.2.11
jaxlib==0.1.64
jsonschema==3.2.0
jupyter-core==4.7.1
kiwisolver==1.3.1
kombu==5.0.2
matplotlib==3.3.4
mpmath==1.2.1
nbformat==5.1.2
networkx==2.5
numpy==1.20.1
openfermion==1.0.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
Pillow==8.1.2
pluggy==0.13.1
prompt-toolkit==3.0.17
protobuf==3.13.0
PubChemPy==1.0.4
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
Qulacs==0.2.0
requests==2.25.1
rsa==4.7.2
scipy==1.6.1
six==1.15.0
sortedcontainers==2.3.0
sympy==1.7.1
tequila-basic==1.0.1
toml==0.10.2
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
